requirements.txt
# 智能招聘数据分析平台 - Render 部署依赖
# 使用预编译包避免构建问题

# Web框架
Flask==3.0.3
dash==2.17.1
dash-bootstrap-components==1.5.0

# 数据处理 (使用预编译版本)
pandas==2.0.3
plotly==5.17.0
numpy==1.24.3

# 数据库连接
PyMySQL==1.1.1

# 生产服务器
gunicorn==22.0.0

# 环境变量管理
python-dotenv==1.0.0

# 导出功能
openpyxl==3.1.2

# 日期处理
python-dateutil==2.8.2

# 缓存支持
Flask-Caching==2.1.0
